Update ES16parser9.py
Need to re-write the serial portion some more.   This one works, but I don't like the time.sleep() pause between serial reads.  I think we need a more aggressive  pass2 inter loop with timer.    If we get normal data in the first pass, it means the timer was to short, and maybe we can auto increment that.